+++ src/test/java/org/I0Itec/zkclient/ZkClientSerializationTest.java
@@ -33,7 +33,7 @@ public class ZkClientSerializationTest {
 
     @Test
     public void testBytes() throws Exception {
-        ZkClient zkClient = new ZkClient(_zk.getZkServerAddress(), 2000, 2000, new BytesPushThroughSerializer());
+        ZkClient zkClient = new ZkClient(_zk.getZkServerAddress(), 2000, 30000, new BytesPushThroughSerializer());
         byte[] bytes = new byte[100];
         new Random().nextBytes(bytes);
         zkClient.createPersistent("/a", bytes);
@@ -43,7 +43,7 @@ public class ZkClientSerializationTest {
 
     @Test
     public void testSerializables() throws Exception {
-        ZkClient zkClient = new ZkClient(_zk.getZkServerAddress(), 2000, 2000, new SerializableSerializer());
+        ZkClient zkClient = new ZkClient(_zk.getZkServerAddress(), 2000, 30000, new SerializableSerializer());
         String data = "hello world";
         zkClient.createPersistent("/a", data);
         String readData = zkClient.readData("/a");